services/tiny-llama-service/requirements.txt
paho-mqtt
mysql-connector-python
torch
transformers
accelerate
